requirements.txt
django~=2.2.10
djangorestframework~=3.12.0
django-choices~=1.7.1
drf-nested-routers~=0.93.3
drf-yasg~=1.20.0